Add support for class item chests.
This is an extremely ugly, hacky way to do it, but the feature
should definitely make it into 1.0 - it requires no commands or
any other complex setting up besides flipping one option in the
config-file.

Upon clicking a class sign in the lobby, MobArena will (if the
config-setting is true) search for chests below the sign, and
below (and including) the block behind the sign, i.e. the block
that a wall sign would be attached to. If a chest is found, the
contents will become the class inventory.

The search can be optimized to be a one-off thing unless the
block changes from being a chest. A todo for 1.0, perhaps.