(Possibly) fix invisible bosses.
Unsurprisingly, MobArena's monster handling is sub-optimal, possibly
due to premature optimization. Before, when e.g. a Creeper boss would
explode, it would not be removed from the bosses collection, and thus
the ability thread would continue on running.

This patch adds a tiny bit of overhead w.r.t. removing dead monsters,
but it fixes the issue with Creeper bosses, and with a bit of luck,
mcMMO will not cause issues anymore (yet to be tested).