Skips permission check if player is Op
On my server I disallow MobArena join commands through negative permissions for all users. The only way to join the arena is to click a special sign.

The CommandSigns plugin can execute commands as though the player is Op, but MobArena would still not execute the command since it ignores the Op status of Player.