Remove the lib/ folder from .gitignore.
The purpose of this is to actually commit to version control the jar(s) necessary to compile. It's not a best practice whatsoever, but if we can commit gradle wrappers, we can commit hard-to-incorporate binary dependencies.

As a nice side effect, it should be possible to run some basic CI tooling on the repository.