Revert to items + armor.
The items + armor approach resulted in issues in the join/leave rewrite, because the armor contents are just a subset of the inventory, while the contents are the full set. This is not the case on servers running 1.8, however, where this approach actually results in a loss of armor items in MobArena. This commit reverts to the old way of doing things.

Fixes #465