Remove set/remove class commands.
Similar to a88556771f0d64955656c770c783f884277e485b.

This specific change may stir the pot a bit, since some people may rely on the set command. However, class chests solve a lot of the problems that the initial command set out to tackle, and we can probably do better in a rewrite of the command if it's missed anyway.

The removed commands render a few methods in ArenaMaster obsolete.