Start the join-interrupt-timer from the spectator area.
Following the changes in 266104d0a1ad319da9ded75a17c89b1581ba3d8d, this commit gathers the leave-then-join logic and handles it in one fell swoop when joining from the spectator area. This means that the timer will start, and when it runs out, the player leaves the current arena (if any), and then joins.

There should be no issues with overlapping timers or anything like that, since the lookup happens when the time is up, not before the timer is started.