Stop MobArena interfering with cancelled teleport events
The mobarena.admin.teleport permission should stop mobarena from
cancelling teleports for players with that permission. The previous
implementation also uncancelled events which were cancelled by other
plugins. The new implementation moves this check to prevent this
behavior. All MobArena check are done regardless of permission, but if
the player has the override permission, the event is not cancelled.

Fixes #515 (but untested due to build issues with Vault dependency)