Check if player is in arena on food level change.
This commit fixes a bug introduced in 88ddbe5ac6ac5aa0f18ade50aa376b25546fb058.

During a food level change event, if the arena is running and the food level lock is enabled, we just need to also make sure that the given player is actually in the arena, otherwise it affects everyone on the server.

Fixes #557